+++ examples/app_fractal_dashboard/bags/uptime.rb
@@ -28,9 +28,9 @@ module Uptime
 
   UPDATE = lambda do |message, model|
     case message
-    in [:uptime, { stdout:, status: 0 }]
+    in [{ type: :system, envelope: :uptime, status: 0, stdout: }]
       [model.with(output: Ractor.make_shareable(stdout.strip), loading: false), nil]
-    in [:uptime, { stderr:, _status: }]
+    in [{ type: :system, envelope: :uptime, stderr: }]
       [model.with(output: Ractor.make_shareable("Error: #{stderr.strip}"), loading: false), nil]
     else
       [model, nil]