+++ test/test_system_batch_message.rb
@@ -1,0 +1,53 @@
+# frozen_string_literal: true
+
+#--
+# SPDX-FileCopyrightText: 2026 Kerrick Long <me@kerricklong.com>
+# SPDX-License-Identifier: LGPL-3.0-or-later
+#++
+
+require "test_helper"
+require "ratatui_ruby/test_helper"
+
+class TestSystemBatchMessage < Minitest::Test
+  include RatatuiRuby::TestHelper
+
+  def test_system_batch_emits_message_not_array
+    messages = []
+    model = Ractor.make_shareable({})
+    view = -> (_m, t) { t.clear }
+
+    update = -> (msg, m) do
+      case msg
+      when RatatuiRuby::Event::Key
+        case msg.code
+        when "s" then [m, RatatuiRuby::Tea::Command.system("echo hello", :build)]
+        when "q" then [m, RatatuiRuby::Tea::Command.exit]
+        else [m, nil]
+        end
+      else
+        messages << msg
+        [m, nil]
+      end
+    end
+
+    require "open3"
+    mock_status = Object.new
+    mock_status.define_singleton_method(:exitstatus) { 0 }
+    Open3.stub(:capture3, ["hello\n", "", mock_status]) do
+      with_test_terminal do
+        inject_key("s")
+        inject_sync
+        inject_key("q")
+        RatatuiRuby::Tea::Runtime.run(model:, view:, update:)
+      end
+    end
+
+    # Should receive Message::System::Batch, not array
+    batch_msg = messages.find { |m| m.is_a?(RatatuiRuby::Tea::Message::System::Batch) }
+    refute_nil batch_msg, "Should receive System::Batch message"
+    assert_equal :build, batch_msg.envelope
+    assert_equal "hello\n", batch_msg.stdout
+    assert_equal "", batch_msg.stderr
+    assert_equal 0, batch_msg.status
+  end
+end